LPS-66853 Move GitUtil and GitExcpetion to portal-impl
