+++ portal-impl/src/com/liferay/portal/tools/GitException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.source.formatter;
+package com.liferay.portal.tools;
 
 import com.liferay.portal.kernel.exception.PortalException;
